backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.16,<4.0
djangorestframework-gis>=1.0,<2.0
django-cors-headers>=4.0,<5.0
django-vectortiles>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
